requirements.txt
binance-futures-connector==4.0.0
certifi==2024.7.4
charset-normalizer==3.3.2
cmdstanpy==1.2.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
holidays==0.55
idna==3.8
importlib_resources==6.4.4
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prophet==1.1.5
pycryptodome==3.20.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
stanio==0.5.1
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
